requirements.txt
# Django framework
django==3.1.4

# Django rest framework
djangorestframework==3.12.2
